Deleting process for Transport Layer
