+++ mix.exs
@@ -8,7 +8,7 @@ defmodule Membrane.RTSP.MixProject do
     [
       app: :membrane_rtsp,
       version: @version,
-      elixir: "~> 1.10",
+      elixir: "~> 1.12",
       elixirc_paths: elixirc_paths(Mix.env()),
       start_permanent: Mix.env() == :prod,
       name: "Membrane RTSP",
